requirement.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Hashing==1.1
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
python-dateutil==2.9.0.post0
six==1.16.0
Werkzeug==3.0.1
zipp==3.18.1
